feat: add navbar animation
