+++ src/components/Hero.jsx
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
-import Navbar from './Navbar';
 import { Canvas } from '@react-three/fiber';
 import { OrbitControls, Sphere, MeshDistortMaterial } from '@react-three/drei';
 
@@ -12,6 +11,7 @@ const Section = styled.div`
   flex-direction: column;
   align-items: center;
 justify-content: space-between;
+
 `
 const Container = styled.div`
 height: 100vh;
@@ -19,6 +19,7 @@ scroll-snap-align: center;
 width: 1400px;
   display: flex;
 justify-content: space-between;
+margin-top: 5rem;
 `
 
 const LeftSection = styled.div`
@@ -146,7 +147,7 @@ export default class Hero extends Component {
         return (
             <Section id="home">
                 
-                <Navbar />
+                
                 <Container>
                     <LeftSection>
                         <Title>
